+++ src/theme/index.ts
@@ -6,6 +6,7 @@ import { spacing } from "./foundations/spacing";
 import breakpoints from "./foundations/breakpoints";
 import Accordion from "./components/accordion";
 import Button from "./components/button";
+import Heading from "./components/heading";
 import Radio from "./components/radio";
 import CustomRadioGroup from "./components/customRadioGroup";
 import Tabs from "./components/tabs";
@@ -41,8 +42,9 @@ const theme = extendTheme({
   components: {
     Accordion,
     Button,
-    CustomRadioGroup,
+    Heading,
     Radio,
+    CustomRadioGroup,
     Tabs,
   },
   // Chakra prefixes its own CSS variables with `--chakra` by default but this