+++ CHANGELOG.md
@@ -8,11 +8,16 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Breaking Changes
+
+- Renames `HeadingDisplaySizes.tsx` to `HeadingTypes.tsx`.
+
 ### Adds
 
 - Integrates Chakra-UI into the codebase.
 - Exports Chakra's `Box`, `Center`, `Circle`, `Square`, `Stack`, `HStack`, and `VStack` components.
 - Composes the DS `Button` component from Chakra's `Button` component.
+- Composes the DS `Heading` component from Chakra's `Heading` component.
 - Composes the DS `Tabs` component from Chakra's `Tabs`, `TabList`, `Tab`, `TabPanels`, and `TabPanel` components.
 - Moved SCSS and CSS variable styles into a custom NYPL theme to extend Chakra's base theme.
 - Updates the `errorText` prop name to `invalidText`.